fix clang build

+++ Drp/stdlogger.h
@@ -30,7 +30,7 @@ static
 void
 log_posix(Logger* logger, LogLevel lvl, const void* data, size_t len){
     int fd = ((int*)(logger+1))[lvl==LOG_PRINT?0:1];
-    int err = write(fd, data, len);
+    ssize_t err = write(fd, data, len);
     (void)err;
 }
 #endif
